gemini/requirements.txt
# Google Gemini AI SDK
google-generativeai>=0.3.0

# Environment variable management
python-dotenv>=1.0.0

# Additional dependencies for text processing
requests>=2.31.0
typing-extensions>=4.8.0
